Move three files
put the generic things in a generic place
