--- src/handlers/commands/RetryDeploy.ts
@@ -1,28 +1,0 @@
-import {HandleCommand, MappedParameter, MappedParameters, Parameter, Secret, Secrets} from "@atomist/automation-client";
-import {Parameters} from "@atomist/automation-client/decorators";
-
-@Parameters()
-export class RetryDeployParameters {
-
-    @Parameter()
-    public repo: string;
-
-    @Parameter()
-    public owner: string;
-
-    @Parameter()
-    public sha: string;
-
-    @Parameter()
-    public branch: string;
-
-    @Parameter()
-    public targetUrl: string;
-
-    @Secret(Secrets.UserToken)
-    public githubToken: string;
-
-    @MappedParameter(MappedParameters.SlackChannelName)
-    public channelName: string;
-
-}
