+++ src/software-delivery-machine/blueprint/deploy/localSpringBootDeployOnSuccessStatus.ts
@@ -15,21 +15,22 @@
  */
 
 import { logger } from "@atomist/automation-client";
-import { DeploySpec, executeDeploy, retryGoal } from "../../../handlers/events/delivery/deploy/executeDeploy";
+import { DeploySpec, executeDeploy } from "../../../handlers/events/delivery/deploy/executeDeploy";
 import { executableJarDeployer } from "../../../handlers/events/delivery/deploy/local/jar/executableJarDeployer";
 import { StartupInfo } from "../../../handlers/events/delivery/deploy/local/LocalDeployerOptions";
 import { mavenDeployer } from "../../../handlers/events/delivery/deploy/local/maven/mavenSourceDeployer";
 import { StagingDeploymentGoal, StagingEndpointGoal } from "../../../handlers/events/delivery/goals/httpServiceGoals";
 import { OnSupersededStatus } from "../../../handlers/events/delivery/superseded/OnSuperseded";
 import { DefaultArtifactStore } from "../artifactStore";
-import { ExecuteGoalOnSuccessStatus } from "../../../handlers/events/delivery/deploy/ExecuteGoalOnSuccessStatus";
+import { ExecuteGoalOnSuccessStatus } from "../../../handlers/events/delivery/ExecuteGoalOnSuccessStatus";
 import {
     ManagedDeploymentTargeter,
     ManagedDeploymentTargetInfo,
     targetInfoForAllBranches
 } from "../../../handlers/events/delivery/deploy/local/appManagement";
 import { Deployer, FunctionalUnit } from "../../../";
-import { ExecuteGoalOnPendingStatus } from "../../../handlers/events/delivery/build/ExecuteGoalOnPendingStatus";
+import { ExecuteGoalOnPendingStatus } from "../../../handlers/events/delivery/ExecuteGoalOnPendingStatus";
+import { retryGoal } from "../../../handlers/commands/RetryGoal";
 
 /**
  * Deploy to the automation client node
